requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-jose[cryptography]
passlib
bcrypt
python-multipart
SQLAlchemy
slowapi